% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecoMetDIA.R
\name{DecoMetDIA}
\alias{DecoMetDIA}
\title{DecoMetDIA: Deconvolution of Multiplexed MS/MS Spectra for Metabolite
    Identification in SWATH-MS based Untargeted Metabolomics}
\usage{
DecoMetDIA(d.in = ".", d.out = "DecoMetDIA_Result", nSlaves = 6,
  peakwidth = c(5, 30), sn = 6, ppm.pd = 25,
  polarity = c("positive", "negative"), lc = c("HILIC", "RP"),
  minfrac = 0.5, bw = 5, mzwid = 0.015, is.ms1.filter = FALSE,
  thr.ms1.filter = NULL, filed.ms1.filter = c("maxo", "into", "intb"),
  is.plot.eic.feature = F, is.check.eic = FALSE,
  is.PlotEIC.MS2 = FALSE, snthr.ms2 = 5,
  fp.swath.setup = "../SWATHsetup.csv", ppm.ms2.mtp = 15,
  int.filter = 50, minfrac.vote = 0.5, is.plot.eic.spec = F,
  is.output.spec.smp = TRUE, is.span.constant = FALSE, rerun = FALSE,
  rerun.ms2 = FALSE, is.ms1.only = FALSE, is.dec.all = FALSE,
  is.dec.smoothed = FALSE, is.norm.consensus = TRUE)
}
\arguments{
\item{d.in}{Working directory, where the well orgnized data are stored}

\item{d.out}{Directory for saving results}

\item{nSlaves}{No. of threads for data processing}

\item{peakwidth}{Chromatographic peak width, given as range (min,max) in seconds
(\code{xcms})}

\item{sn}{Signal to noise ratio cutoff for peak detection using \code{xcms}}

\item{ppm.pd}{Maxmial tolerated m/z deviation in consecutive scans, in ppm
for MS1 peak detection}

\item{polarity}{Polarity setup for ionization}

\item{lc}{LC column used for chromatographic seperation}

\item{minfrac}{minimum fraction of samples necessary in at least one of the
sample groups for it to be a valid group when group peaks with 'denisty'
method in \code{xcms}}

\item{bw}{bandwidth (standard deviation or half width at half maximum)
of gaussian smoothing kernel to apply to the peak density chromatogram when
group peaks with 'denisty' method in \code{xcms}}

\item{mzwid}{width of overlapping m/z slices to use for creating peak density
chromatograms and grouping peaks across samples when group peaks with
'denisty' method in \code{xcms}}

\item{is.ms1.filter}{logical value for indicating if filter isotopic or low
abundant peaks}

\item{thr.ms1.filter}{MS1 peak filter threshold for filtering low abundant peaks}

\item{filed.ms1.filter}{MS1 peak filter method (maxo:intensity, into:raw peak area or
intb:base-line corrected peak area)}

\item{is.plot.eic.feature}{logical value for indicating if plot EICs of features}

\item{is.check.eic}{logical value for indicating if plot peak boundaries
when plotting feature EICs}

\item{is.PlotEIC.MS2}{logical value for indicating if plot EICs of fragment ions}

\item{snthr.ms2}{signal to noise threshold for MS2 peak detection}

\item{fp.swath.setup}{file path for SWATH setup file (.csv format recording swath windows)}

\item{ppm.ms2.mtp}{ppm threshold for EIC extraction of fragment ions in multiplexed spectrum}

\item{int.filter}{intensity threshold for removing low abundant fragment ions
in multiplexed spectrum}

\item{minfrac.vote}{= 0.5,}

\item{is.output.spec.smp}{logical value for indicating if output deconvoluted spectra
for each sample}

\item{is.span.constant}{logical value for indicating if using constant span
when smoothing EICs}

\item{rerun}{logical value for indicating if rerun the already processed peak
detection results}

\item{rerun.ms2}{logical value for indicating if rerun the already prcoessed
MS2 data analysis process}

\item{is.ms1.only}{logical value for indicating if only detect peaks}

\item{is.dec.all}{logical value for indicating if deconvolute all EICs, if not,
the 'simple' peaks are ignored}

\item{is.dec.smoothed}{logical value for indicating if using smoothed peaks as
model peaks.}

\item{is.norm.consensus}{logical value for indicating if normalize intensities
to 1 when consensusing spectra cross samples.}
}
\description{
DecoMetDIA picks peaks from SWATH acquired data with XCMS and
    extracts MS2 EICs under related SWATH window for each feature of different
    sample. Deconvolution algorithms are applied to MS2 EICs for each feature
    and thus MS2 spectra are generated. The features are identified by matching
    the obtained spectra with librarial ones.
}
